split reducers into separate files
